+++ src/pages/Comidas.js
@@ -1,35 +1,22 @@
 import React, { useContext } from 'react';
-import { Redirect } from 'react-router-dom';
 import propTypes from 'prop-types';
 import Header from '../components/Header';
 import Footer from '../components/Footer';
 import { RecipesContext } from '../context/Recipes';
-import RecipesList from '../components/RecipesList';
+import ShowRecipes from '../components/ShowRecipes';
 import CategoriesList from '../components/CategoriesList';
 import useFetchRandom from '../hooks/useFetchRandom';
 import './MainPage.css';
 
-const oneRecipe = (fetchResult) => {
-  const { idMeal, idDrink } = fetchResult[0];
-  const idRecipe = idMeal || idDrink;
-  return <Redirect to={`./receita/${idRecipe}`} />;
-};
-
-const showRecipes = (fetchResult = null) => {
-  if (fetchResult === null) return <h2>Nada encontrado.</h2>;
-  if (fetchResult.length > 1) return <RecipesList />;
-  return oneRecipe(fetchResult);
-};
-
 const Comidas = ({ match }) => {
-  const { isFetching, fetchResult } = useContext(RecipesContext);
+  const { isFetching } = useContext(RecipesContext);
   const [data] = useFetchRandom(match);
   return (
     <div>
       <Header />
       <CategoriesList random={data} match={match} />
       <div className="MainContainerPage">
-        {isFetching ? <h2>Buscando...</h2> : showRecipes(fetchResult)}
+        {isFetching ? <h2>Buscando...</h2> : <ShowRecipes />}
       </div>
       <Footer />
     </div>
